fix: move verification to spec level
