+++ tests/acceptance/tests/Categories/AddLandingPageInCategories.spec.ts
@@ -5,7 +5,7 @@ test('Shop administrator should be able to create a landing page.', {tag: '@Cate
     ShopAdmin,
     IdProvider,
     TestDataService,
-    AdminCategories, CreateLandingPage, VerifyLandingPage, AdminApiContext,
+    AdminCategories, CreateLandingPage, AdminLandingPageDetail, AdminApiContext,
 }) => {
     const layoutUuid = IdProvider.getIdPair().uuid;
     const layoutName = `test ${layoutUuid}`;
@@ -14,6 +14,7 @@ test('Shop administrator should be able to create a landing page.', {tag: '@Cate
         name: `Landing Page ${IdProvider.getIdPair().uuid}`,
         salesChannel: 'Storefront',
         seoUrl: `landing-${IdProvider.getIdPair().id}`,
+        status: true,
     };
 
     await test.step('Create a landing page layout via API.', async () => {
@@ -32,7 +33,25 @@ test('Shop administrator should be able to create a landing page.', {tag: '@Cate
     });
 
     await test.step('Verify a new landing page created and assigned layout.', async () => {
-        await ShopAdmin.attemptsTo(VerifyLandingPage(layoutName, landingPageData, true));
+        // Verify created landing page
+        const createdLandingPage = AdminCategories.landingPageItems.locator(`text="${landingPageData.name}"`);
+        await createdLandingPage.click();
+        await AdminLandingPageDetail.loadingSpinner.waitFor({ state: 'hidden' });
+
+        // Verify general tab detail
+        await ShopAdmin.expects(AdminLandingPageDetail.nameInput).toHaveValue(landingPageData.name);
+        await ShopAdmin.expects(AdminLandingPageDetail.landingPageStatus).toBeChecked({ checked: landingPageData.status });
+        await ShopAdmin.expects(AdminLandingPageDetail.salesChannelSelectionList).toHaveText(landingPageData.salesChannel);
+        await ShopAdmin.expects(AdminLandingPageDetail.seoUrlInput).toHaveValue(landingPageData.seoUrl);
+        // Verify layout tab detail
+        await AdminLandingPageDetail.layoutTab.click();
+        await ShopAdmin.expects(AdminLandingPageDetail.layoutAssignmentCardTitle).toHaveText(layoutName);
+        await ShopAdmin.expects(AdminLandingPageDetail.layoutAssignmentCardHeadline).toHaveText(layoutName);
+
+        await ShopAdmin.expects(AdminLandingPageDetail.layoutAssignmentContentSection).toBeVisible();
+        await ShopAdmin.expects(AdminLandingPageDetail.layoutResetButton).toBeVisible();
+        await ShopAdmin.expects(AdminLandingPageDetail.changeLayoutButton).toBeVisible();
+        await ShopAdmin.expects(AdminLandingPageDetail.editInDesignerButton).toBeVisible();
     });
 
     await test.step('Clean up the created landing page via API.', async () => {
